Place datatypes_conf.xml in the main galaxy directory and overwrite the old one.
Place images.py in lib/galaxy/datatypes and overwrite the old one

Note: Galaxy only recognizes lower case file extensions.  You may need to
manually indicate the file type.  To do so, upload the file and then click on
pencil icon and the the "datatypes tag." Click on the bar and options should
drop down for you to use.